apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
cryptography==43.0.1
celery[redis]==5.4.0
httpx==0.27.2
anthropic==0.34.2
python-dotenv==1.0.1
pydantic-settings==2.5.2
pydantic==2.9.2
